// src/types/IDataItem.ts
export interface IDataItem {
    address: string;
    buildings_disrepair: number;
    buildings_management: number;
    buildings_repair: number;
    computers: number;
    copies: number;
    copies_electronic: number;
    copies_issued: number;
    copies_issued_children: number;
    digital_catalogs: number;
    dropped_copies: number;
    electronic_catalogue_volume: number;
    employees: number;
    employees_staff: number;
    formname: string;
    fullname: string;
    funds: number;
    funds_acquisition: number;
    funds_budget: number;
    funds_entrepreneurial: number;
    funds_used: number;
    internet: number;
    internet_catalogs: number;
    internet_catalogue_volume: number;
    issued_electronic: number;
    kopuk: string;
    libraries: number;
    libraries_computers: number;
    order: number;
    out_of_instances: number;
    period: string;
    received_copies: number;
    received_electronic: number;
    site: number;
    staff_higheeducated: number;
    subscribers: number;
    territory: string;
    users: number;
    users_children: number;
    visits: number;
    visits_sites: number;
}